[007] rename store
