+++ src/store/index.js
@@ -1,14 +1,14 @@
 import Vue from "vue";
 import Vuex from "vuex";
 
-import routingError from "./routingError";
+import routingErrors from "./routingErrors";
 import popularList from "./popularList";
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
     modules: {
-        routingError,
+        routingErrors,
 
         popularList,
     },
